+++ source/xq/services/encrypt_file.c
@@ -392,8 +392,8 @@ _Bool xq_encrypt_file_start(struct xq_config *config,  const char *in_file_path,
   return 0;
 }
 
-int xq_encrypt_file_step(struct xq_file_stream *stream_info, uint8_t *data,
-                           int data_length, struct xq_error_info *error) {
+size_t xq_encrypt_file_step(struct xq_file_stream *stream_info, uint8_t *data,
+                           size_t data_length, struct xq_error_info *error) {
 
   return xq_otp_encrypt_file_step(stream_info, data, data_length, error);
 }
